Renames gcs/errors to gcs/gcserr
Resolves: #26
