+++ service/gcs/bridge/bridge.go
@@ -11,7 +11,7 @@ import (
 	"sync"
 
 	"github.com/Microsoft/opengcs/service/gcs/core"
-	gcserr "github.com/Microsoft/opengcs/service/gcs/errors"
+	"github.com/Microsoft/opengcs/service/gcs/gcserr"
 	"github.com/Microsoft/opengcs/service/gcs/oslayer"
 	"github.com/Microsoft/opengcs/service/gcs/prot"
 	"github.com/Microsoft/opengcs/service/gcs/transport"
