+++ service/gcs/gcserr/errors_suite_test.go
@@ -1,12 +1,12 @@
-package errors
+package gcserr
 
 import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/sirupsen/logrus"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/sirupsen/logrus"
 )
 
 func TestErrors(t *testing.T) {
